Add files via upload
So here i have used a new label rank of the person , which labels the person as per his designation i.e. Sir , Royalty , Master etc. Age of the missing variables also have been mean of the rank , gender and title 

Coming to the model its Randomforest with GridsearchCV to find optimal parameters.